Create README.md
Hi!

Toolbox is (as its name says) a compilation of tiny tools, designs and some other random stuff that maybe weren't big enough as to dedicate an entire website to, but important enough as to have a place inside Toolbox.  

As a personal meaning, this was made to practice and show my current web dev abilities such as basic JS functions, Consuming APIs, CSS3 Styling and such.